Reorganized query interfaces and migrated namespaces from `Domain` to `Application`. Fixed typos in docblocks, replacing "modal" with "model".
