+++ src/User/Identity/Application/Query/User/FindUserByEmailInterface.php
@@ -11,7 +11,7 @@
  */
 declare(strict_types=1);
 
-namespace Webify\User\Identity\Domain\Query\User;
+namespace Webify\User\Identity\Application\Query\User;
 
 use Webify\User\Identity\Domain\ValueObject\UserEmail;
 
@@ -25,7 +25,7 @@ interface FindUserByEmailInterface
 	 *
 	 * @param UserEmail $email the email of the user to find
 	 *
-	 * @return null|User the User read-modal if found, or null if no user matches the given email
+	 * @return null|User the User read-model if found, or null if no user matches the given email
 	 */
 	public function find(UserEmail $email): ?User;
 }
